+++ tests/unit/call.ts
@@ -1119,7 +1119,7 @@ describe('Call', function() {
     beforeEach(() => {
       message = {
         content: { foo: 'foo' },
-        messageType: Call.MessageType.UserDefinedMessage,
+        messageType: 'user-defined-message',
       };
     });
 
@@ -1135,8 +1135,7 @@ describe('Call', function() {
           assert.throws(
             () => conn.sendMessage({ ...message, messageType }),
             new InvalidArgumentError(
-              '`messageType` must be an enumeration value of ' +
-              '`Call.MessageType` or a string.',
+              '`messageType` must be a string.',
             ),
           );
         });
@@ -1185,7 +1184,7 @@ describe('Call', function() {
         mockCallSid,
         message.content,
         undefined,
-        Call.MessageType.UserDefinedMessage,
+        'user-defined-message',
         'foobar-voice-event-sid',
       );
     });
@@ -2007,7 +2006,7 @@ describe('Call', function() {
             foo: 'bar',
           },
           contentType: 'application/json',
-          messageType: 'foo-bar',
+          messageType: 'user-defined-message',
         };
 
         mockAckPayload = {
@@ -2030,6 +2029,18 @@ describe('Call', function() {
         });
         pstream.emit('ack', { ...mockAckPayload, voiceeventsid: sid });
         assert.deepEqual(await payloadPromise, { ...mockMessagePayload, voiceEventSid: sid });
+      });
+
+      it('should publish a user-defined-message sent event', () => {
+        const sid = conn.sendMessage(mockMessagePayload);
+        const payloadPromise = new Promise((resolve) => {
+          conn.on('messageSent', resolve);
+        });
+        pstream.emit('ack', { ...mockAckPayload, voiceeventsid: sid });
+        payloadPromise.then((payload) => {
+          sinon.assert.calledWith(publisher.info, 'user-defined-message', 'sent');
+          assert.deepEqual(payload, { ...mockMessagePayload, voiceEventSid: sid });
+        });
       });
 
       it('should ignore ack when callSids do not match', async () => {
@@ -2088,7 +2099,14 @@ describe('Call', function() {
         const payloadPromise = new Promise((resolve, reject) => {
           conn.on('messageSent', reject);
         });
-        pstream.emit('error', { callsid: mockcallsid, voiceeventsid: sid });
+        pstream.emit('error', {
+          callsid: mockcallsid,
+          voiceeventsid: sid,
+          error: {
+            code: 123,
+            message: 'foo',
+          }
+        });
         pstream.emit('ack', { ...mockAckPayload, voiceeventsid: sid });
         await Promise.race([
           wait(1),
@@ -2170,6 +2188,44 @@ describe('Call', function() {
           wait(1),
           messagePromise,
         ]);
+      });
+
+      it('should publish a user-defined-message received event', () => {
+        const payloadPromise = new Promise((resolve) => {
+          conn.on('messageReceived', resolve);
+        });
+        pstream.emit('message', mockPayload);
+        payloadPromise.then((payload) => {
+          sinon.assert.calledWith(publisher.info, 'user-defined-message', 'received');
+          assert.deepEqual(payload, {
+            content: mockPayload.content,
+            contentType: mockPayload.contenttype,
+            messageType: mockPayload.messagetype,
+            voiceEventSid: mockPayload.voiceeventsid,
+          });
+        });
+      });
+
+      it('should publish a user-defined-message error event', () => {
+        const payloadPromise = new Promise((resolve, reject) => {
+          conn.on('error', reject);
+        });
+        pstream.emit('error', {
+          callsid: mockcallsid,
+          voiceeventsid: 'mock-voiceeventsid-foobar',
+          error: {
+            code: 111,
+            message: 'foo-error',
+          }
+        });
+        payloadPromise.then((payload) => {
+          sinon.assert.calledWith(publisher.error, 'user-defined-message', 'error');
+          assert.deepEqual(payload, {
+            code: 111,
+            message: 'foo-error',
+            voiceEventSid: 'mock-voiceeventsid-foobar',
+          });
+        });
       });
     });
   });